requirements.txt
certifi==2018.1.18
chardet==3.0.4
gevent==1.2.2
greenlet==0.4.13
grequests==0.3.0
idna==2.6
python-firebase==1.2
requests==2.18.4
urllib3==1.22
